P1-76, P1-79 fixes

+++ frontend/src/api/userIdentity.js
@@ -1,20 +1,25 @@
 
 function saveUserSession(context, userName, userType) {
     return new Promise((resolve, reject) => {
-        context.setUserName(userName);//set user context data
-        context.setUserType(userType);
-        sessionStorage.setItem("userName", userName);// set session storage data (user name, user type)
-        sessionStorage.setItem("userType", userType);
-        resolve("ok");
+        if(context) {
+            context.setUserName(userName);//set user context data
+            context.setUserType(userType);
+            sessionStorage.setItem("userName", userName);// set session storage data (user name, user type)
+            sessionStorage.setItem("userType", userType);
+            resolve("ok");
+            }
     });
 }
 
 function removeUserSession(context) {
     return new Promise((resolve, reject) => {
-        context.setUserName(null);//set user context data
-        context.setUserType(null);
-        sessionStorage.clear();//delete session storage data
-        resolve("ok");
+        if(context) {
+            context.setUserName(null);//set user context data
+            context.setUserType(null);
+            sessionStorage.clear();//delete session storage data
+            resolve("ok");
+        }
+        
     });
 }
